+++ dbt/include/clickhouse/macros/materializations/materialized_view.sql
@@ -7,6 +7,8 @@
 
   {%- set target_relation = this.incorporate(type='table') -%}
   {%- set cluster_clause = on_cluster_clause(target_relation) -%}
+  {%- set refreshable_clause = refreshable_mv_clause() -%}
+
 
   {# look for an existing relation for the target table and create backup relations if necessary #}
   {%- set existing_relation = load_cached_relation(this) -%}
@@ -35,13 +37,13 @@
   {{ run_hooks(pre_hooks, inside_transaction=True) }}
 
   -- extract the names of the materialized views from the sql
-  {% set view_names = modules.re.findall('--([^:]+):begin', sql) %}
+  {% set view_names = modules.re.findall('--(?:\s)?([^:]+):begin', sql) %}
 
   -- extract the sql for each of the materialized view into a map
   {% set views = {} %}
   {% if view_names %}
     {% for view_name in view_names %}
-      {% set view_sql = modules.re.findall('--' + view_name + ':begin(.*)--' + view_name + ':end', sql, flags=modules.re.DOTALL)[0] %}
+      {% set view_sql = modules.re.findall('--(?:\s)?' + view_name + ':begin(.*)--(?:\s)?' + view_name + ':end', sql, flags=modules.re.DOTALL)[0] %}
       {%- set _ = views.update({view_name: view_sql}) -%}
     {% endfor %}
   {% else %}
@@ -50,7 +52,8 @@
 
   {% if backup_relation is none %}
     {{ log('Creating new materialized view ' + target_relation.name )}}
-    {{ clickhouse__get_create_materialized_view_as_sql(target_relation, sql, views) }}
+    {% set catchup_data = config.get("catchup", True) %}
+    {{ clickhouse__get_create_materialized_view_as_sql(target_relation, sql, views, catchup_data) }}
   {% elif existing_relation.can_exchange %}
     {{ log('Replacing existing materialized view ' + target_relation.name) }}
     -- in this section, we look for mvs that has the same pattern as this model, but for some reason,
@@ -68,7 +71,7 @@
         {% endset %}
 
     {% set tables_result = run_query(tables_query) %}
-    {% if tables_result is not none %}
+    {% if tables_result is not none and tables_result.columns %}
         {% set tables = tables_result.columns[0].values() %}
         {{ log('Current mvs found in ClickHouse are: ' + tables | join(', ')) }}
         {% set mv_names = [] %}
@@ -138,13 +141,20 @@
   2. Create a materialized view using the SQL in the model that inserts
   data into the table creating during step 1
 #}
-{% macro clickhouse__get_create_materialized_view_as_sql(relation, sql, views) -%}
+{% macro clickhouse__get_create_materialized_view_as_sql(relation, sql, views, catchup=True ) -%}
   {% call statement('main') %}
+  {% if catchup == True %}
     {{ get_create_table_as_sql(False, relation, sql) }}
+  {% else %}
+  {{ log('Catchup data config was set to false, skipping mv-target-table initial insertion ')}}
+   {% set has_contract = config.get('contract').enforced %}
+    {{ create_table_or_empty(False, relation, sql, has_contract) }}
+  {% endif %}
   {% endcall %}
   {%- set cluster_clause = on_cluster_clause(relation) -%}
+  {%- set refreshable_clause = refreshable_mv_clause() -%}
   {%- set mv_relation = relation.derivative('_mv', 'materialized_view') -%}
-  {{ clickhouse__create_mvs(relation, cluster_clause, views) }}
+  {{ clickhouse__create_mvs(relation, cluster_clause, refreshable_clause, views) }}
 {%- endmacro %}
 
 {% macro clickhouse__drop_mv(mv_relation, cluster_clause)  -%}
@@ -175,6 +185,7 @@
     {{ clickhouse__drop_mv(mv_relation, cluster_clause) }};
     {{ clickhouse__create_mv(mv_relation, target_relation, cluster_clause, view_sql) }};
   {% endif %}
+
 {%- endmacro %}
 
 {% macro clickhouse__drop_mvs(target_relation, cluster_clause, views)  -%}
@@ -201,6 +212,7 @@
 {% macro clickhouse__replace_mv(target_relation, existing_relation, intermediate_relation, backup_relation, sql, views) %}
   {# drop existing materialized view while we recreate the target table #}
   {%- set cluster_clause = on_cluster_clause(target_relation) -%}
+  {%- set refreshable_clause = refreshable_mv_clause() -%}
   {{ clickhouse__drop_mvs(target_relation, cluster_clause, views) }}
 
   {# recreate the target table #}
@@ -211,6 +223,111 @@
   {{ adapter.rename_relation(intermediate_relation, target_relation) }}
 
   {# now that the target table is recreated, we can finally create our new view #}
-  {{ clickhouse__create_mvs(target_relation, cluster_clause, views) }}
+  {{ clickhouse__create_mvs(target_relation, cluster_clause, refreshable_clause, views) }}
 {% endmacro %}
+
+{% macro refreshable_mv_clause() %}
+  {%- if config.get('refreshable') is not none -%}
+
+    {% set refreshable_config = config.get('refreshable') %}
+    {% if refreshable_config is not mapping %}
+      {% do exceptions.raise_compiler_error(
+        "The 'refreshable' configuration must be defined as a dictionary. Please review the docs for more details."
+      ) %}
+    {% endif %}
+
+    {% set refresh_interval = refreshable_config.get('interval', none) %}
+    {% set refresh_randomize = refreshable_config.get('randomize', none) %}
+    {% set depends_on = refreshable_config.get('depends_on', none) %}
+    {% set depends_on_validation = refreshable_config.get('depends_on_validation', false) %}
+    {% set append = refreshable_config.get('append', false) %}
+
+    {% if not refresh_interval %}
+      {% do exceptions.raise_compiler_error(
+        "The 'refreshable' configuration is defined, but 'interval' is missing. "
+        ~ "This is required to create a refreshable materialized view."
+      ) %}
+    {% endif %}
+
+    {% if refresh_interval %}
+      REFRESH {{ refresh_interval }}
+      {# This is a comment to force a new line between REFRESH and RANDOMIZE clauses #}
+      {%- if refresh_randomize -%}
+        RANDOMIZE FOR {{ refresh_randomize }}
+      {%- endif -%}
+    {% endif %}
+
+    {% if depends_on %}
+      {% set depends_on_list = [] %}
+
+      {% if depends_on is string %}
+        {% set depends_on_list = [depends_on] %}
+      {% elif depends_on is iterable %}
+        {% set temp_list = depends_on_list %}
+        {%- for dep in depends_on %}
+          {% if dep is string %}
+            {% do temp_list.append(dep) %}
+          {% else %}
+            {% do exceptions.raise_compiler_error(
+              "The 'depends_on' configuration must be either a string or a list of strings."
+            ) %}
+          {% endif %}
+        {% endfor %}
+        {% set depends_on_list = temp_list %}
+      {% else %}
+        {% do exceptions.raise_compiler_error(
+          "The 'depends_on' configuration must be either a string or a list of strings."
+        ) %}
+      {% endif %}
+
+      {% if depends_on_validation and depends_on_list | length > 0 %}
+        {%- for dep in depends_on_list %}
+          {% do validate_refreshable_mv_existence(dep) %}
+        {%- endfor %}
+      {% endif %}
+
+      DEPENDS ON {{ depends_on_list | join(', ') }}
+    {% endif %}
+
+    {%- if append -%}
+      APPEND
+    {%- endif -%}
+
+  {%- endif -%}
+{% endmacro %}
+
+
+{% macro validate_refreshable_mv_existence(mv) %}
+  {{ log(mv + ' was recognized as a refreshable mv dependency, checking its existence') }}
+  {% set default_database = "default" %}
+
+  {%- set parts = mv.split('.') %}
+  {%- if parts | length == 2 %}
+    {%- set database = parts[0] %}
+    {%- set table = parts[1] %}
+  {%- else %}
+    {%- set database = default_database %}
+    {%- set table = parts[0] %}
+  {%- endif %}
+
+  {%- set condition = "database='" + database + "' and view='" + table + "'" %}
+
+  {% set query %}
+    select count(*)
+    from system.view_refreshes
+    where {{ condition }}
+  {% endset %}
+
+  {% set tables_result = run_query(query) %}
+    {{ log(tables_result.columns[0].values()[0]) }}
+  {% if tables_result.columns[0].values()[0] > 0 %}
+    {{ log('MV ' + mv + ' exists.') }}
+  {% else %}
+    {% do exceptions.raise_compiler_error(
+      'No existing MV found matching MV: ' + mv
+    ) %}
+  {% endif %}
+{% endmacro %}
+
+